feat(ProductVariantSelectorBlock): added a new block
